submission_helper/requirements.txt
altgraph
certifi
charset-normalizer
docutils
idna
importlib-metadata
infinity
Kivy
Kivy-Garden
packaging
pillow
Pygments
pyinstaller==5.0
pyinstaller-hooks-contrib
python-intervals
requests
screeninfo
urllib3
zipp
